+++ src/Lykke.Service.Assets/Controllers/v2/AssetAttributesController.cs
@@ -10,6 +10,7 @@ using Swashbuckle.AspNetCore.SwaggerGen;
 
 namespace Lykke.Service.Assets.Controllers.V2
 {
+    [ApiController]
     [Route("api/v2/asset-attributes")]
     public class AssetAttributesController : Controller
     {
