+++ src/Lykke.Service.Assets/Controllers/v2/AssetsController.cs
@@ -11,6 +11,7 @@ using System.Threading.Tasks;
 
 namespace Lykke.Service.Assets.Controllers.V2
 {
+    [ApiController]
     [Route("api/v2/assets")]
     public class AssetsController : Controller
     {
