+++ src/Lykke.Service.Assets/Modules/CacheModule.cs
@@ -65,6 +65,7 @@ namespace Lykke.Service.Assets.Modules
             System.Threading.ThreadPool.SetMinThreads(300, 300);
             var options = ConfigurationOptions.Parse(_settings.RedisSettings.Configuration);
             options.ReconnectRetryPolicy = new ExponentialRetry(3000, 15000);
+            options.ClientName = "Lykke.Service.Assets";
 
             var redis = ConnectionMultiplexer.Connect(options);
