+++ src/Lykke.Service.Assets/Settings/AssetsSettings.cs
@@ -1,0 +1,13 @@
+﻿namespace Lykke.Service.Assets.Settings
+{
+    public partial class ApplicationSettings
+    {
+        public class AssetsSettings
+        {
+            public DbSettings Db { get; set; }
+            public RabbitSettings Rabbit { get; set; }
+            public RedisSettings RedisSettings { get; set; }
+            public CqrsSettings Cqrs { get; set; }
+        }
+    }
+}
